+++ Spendr/Models/Entry.swift
@@ -17,7 +17,7 @@ struct Entry: Identifiable {
     let id = UUID()
     var type: EntryType
     var date: Date
-    var amount: Float
+    var amount: Double
     var category: Category
     var name: String
     var account: String
